--- src/components/index.tsx
@@ -1,2 +1,0 @@
-export * from "./TagCard";
-export * from "./ResultBar";
